pyproject.toml
[tool.poetry]
name = "django-sample-project"
version = "0.1.0"
description = ""
authors = ["kyeon06 <jkyeon06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
django-allauth = "^0.59.0"
dj-rest-auth = "^5.0.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
isort = "^5.13.2"
black = "^23.12.0"
flake8 = "^6.1.0"

[tool.black]
line-length = 140
target-version = ['py310']
force-exclude = 'migrations'

[tool.isort]
profile = "black"
filter_files = true
skip_glob = ["*/migrations/*", "config/*", ".gitignore"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
